Uncaught exception should fail a passed test
Commit 4067c0658e3e354b74a3614cfc4e50e4a8735512 addresses the case where a test
fails and a second uncaught exception is thrown. I believe if a test passes and
an uncaught exception is thrown, it should have its status changed to failed,
since this is now the first exception that's been thrown in the test.

One example: We try to prevent tests from making unexpected HTTP requests. Some
of these are made in the next tick, e.g. after a test has returned and has been
marked as complete. These are currently marked as passing, e.g. the failure is
ignored. I believe they should be marked as failed.